Extract SubtypingHelper for unit test
