+++ test/test_helper.rb
@@ -270,3 +270,99 @@ module ASTAssertion
     yield name: annot.name, type: annot.type if block_given?
   end
 end
+
+module SubtypingHelper
+  BUILTIN = <<-EOS
+class BasicObject
+end
+
+class Object <: BasicObject
+  def class: -> module
+  def tap: { (instance) -> any } -> instance
+end
+
+class Class<'a>
+end
+
+class Module
+  def block_given?: -> any
+end
+
+class String
+  def to_str: -> String
+  def +: (String) -> String
+end
+
+class Integer
+  def to_int: -> Integer
+end
+
+class Range<'a>
+  def begin: -> 'a
+  def end: -> 'a
+end
+
+class Regexp
+end
+
+class Array<'a>
+  def []: (Integer) -> 'a
+  def []=: (Integer, 'a) -> 'a
+  def <<: ('a) -> self
+  def each: { ('a) -> any } -> self
+  def zip: <'b> (Array<'b>) -> Array<'a | 'b>
+  def each_with_object: <'b> ('b) { ('a, 'b) -> any } -> 'b
+end
+  EOS
+
+  DEFAULT_SIGS = <<-EOS
+interface _A
+  def +: (_A) -> _A
+end
+
+interface _B
+end
+
+interface _C
+  def f: () -> _A
+  def g: (_A, ?_B) -> _B
+  def h: (a: _A, ?b: _B) -> _C
+end
+
+interface _D
+  def foo: () -> any
+end
+
+interface _X
+  def f: () { (_A) -> _D } -> _C 
+end
+
+interface _Kernel
+  def foo: (_A) -> _B
+         | (_C) -> _D
+end
+
+interface _PolyMethod
+  def snd: <'a>(any, 'a) -> 'a
+  def try: <'a> { (any) -> 'a } -> 'a
+end
+
+module Foo<'a>
+end
+  EOS
+
+  def new_subtyping_checker(sigs = DEFAULT_SIGS)
+    signatures = Steep::AST::Signature::Env.new.tap do |env|
+      parse_signature(BUILTIN).each do |sig|
+        env.add sig
+      end
+
+      parse_signature(sigs).each do |sig|
+        env.add sig
+      end
+    end
+
+    builder = Steep::Interface::Builder.new(signatures: signatures)
+    Steep::Subtyping::Check.new(builder: builder)
+  end
+end
